+++ cmd/resim/commands/sync/apply.go
@@ -129,7 +129,7 @@ func updateSingleExperience(
 			ContainerTimeoutSeconds: update.New.ContainerTimeoutSeconds,
 			Profile:                 update.New.Profile,
 			EnvironmentVariables:    update.New.EnvironmentVariables,
-			CacheExempt:             &update.New.CacheExempt,
+			CacheExempt:             update.New.CacheExempt,
 		}
 
 		response, err := client.CreateExperienceWithResponse(context.Background(), projectID, body)
@@ -143,7 +143,7 @@ func updateSingleExperience(
 
 		// Update the match with then new experience ID. This way the other updates (which
 		// reference this by pointer) will have the ID.
-		update.New.ExperienceID = &ExperienceIDWrapper{ID: response.JSON201.ExperienceID}
+		update.New.ExperienceID = &response.JSON201.ExperienceID
 		return nil
 	}
 	if update.New.ExperienceID == nil {
@@ -156,7 +156,7 @@ func updateSingleExperience(
 		return nil
 	}
 	// Update
-	experienceID := update.New.ExperienceID.ID
+	experienceID := *update.New.ExperienceID
 	if update.Original.Archived {
 		// Restore
 		response, err := client.RestoreExperienceWithResponse(context.Background(), projectID, experienceID)
@@ -168,7 +168,7 @@ func updateSingleExperience(
 			return err
 		}
 	}
-	updateMask := []string{"name", "description", "cacheExempt", "locations"}
+	updateMask := []string{"name", "description", "locations"}
 
 	// These are only updated if they're included. Otherwise they retain their current
 	// value. Users should probably include them anyway.
@@ -181,6 +181,9 @@ func updateSingleExperience(
 	if update.New.EnvironmentVariables != nil {
 		updateMask = append(updateMask, "environmentVariables")
 	}
+	if update.New.CacheExempt != nil {
+		updateMask = append(updateMask, "cacheExempt")
+	}
 
 	body := api.UpdateExperienceInput{
 		Experience: &api.UpdateExperienceFields{
@@ -190,7 +193,7 @@ func updateSingleExperience(
 			ContainerTimeoutSeconds: update.New.ContainerTimeoutSeconds,
 			Profile:                 update.New.Profile,
 			EnvironmentVariables:    update.New.EnvironmentVariables,
-			CacheExempt:             &update.New.CacheExempt,
+			CacheExempt:             update.New.CacheExempt,
 		},
 		UpdateMask: &updateMask,
 	}
@@ -217,7 +220,7 @@ func maybeArchiveExperiences(
 			// Fatal since this should *never* happen. It's a bug in the api client if so.
 			return fmt.Errorf("Trying to archive with unset experience ID")
 		}
-		experiencesToArchive = append(experiencesToArchive, update.New.ExperienceID.ID)
+		experiencesToArchive = append(experiencesToArchive, *update.New.ExperienceID)
 	}
 	if len(experiencesToArchive) == 0 {
 		return nil
@@ -266,7 +269,7 @@ func updateSingleTag(
 			if e.ExperienceID == nil {
 				return fmt.Errorf("Experience has no ID. Maybe we failed to create it? %s", e.Name)
 			}
-			experienceIDs = append(experienceIDs, e.ExperienceID.ID)
+			experienceIDs = append(experienceIDs, *e.ExperienceID)
 		}
 
 		body := api.AddTagsToExperiencesInput{
@@ -290,7 +293,7 @@ func updateSingleTag(
 		go func() {
 			if err := removeTagFromExperience(client, projectID,
 				updates.TagID,
-				removal.ExperienceID.ID); err != nil {
+				*removal.ExperienceID); err != nil {
 				errCh <- err
 			}
 			wg.Done()
@@ -320,7 +323,7 @@ func updateSingleSystem(
 			if e.ExperienceID == nil {
 				return fmt.Errorf("Experience has no ID. Maybe we failed to create it? %s", e.Name)
 			}
-			experienceIDs = append(experienceIDs, e.ExperienceID.ID)
+			experienceIDs = append(experienceIDs, *e.ExperienceID)
 		}
 		body := api.MutateSystemsToExperienceInput{
 			SystemIDs:   []SystemID{updates.SystemID},
@@ -346,7 +349,7 @@ func updateSingleTestSuite(
 	experiences := []ExperienceID{}
 
 	for _, exp := range update.Experiences {
-		experiences = append(experiences, exp.ExperienceID.ID)
+		experiences = append(experiences, *exp.ExperienceID)
 	}
 
 	body := api.ReviseTestSuiteInput{