+++ cmd/resim/commands/sync/command.go
@@ -54,10 +54,10 @@ func CloneExperiences(client api.ClientWithResponsesInterface,
 	if err != nil {
 		log.Fatalf("%v", err)
 	}
-	config.Experiences = []*Experience{}
+	config.Experiences = []Experience{}
 	for _, experience := range currentState.ExperiencesByName {
 		if !experience.Archived {
-			config.Experiences = append(config.Experiences, experience)
+			config.Experiences = append(config.Experiences, *experience)
 		}
 	}
 	writeConfigToFile(config, configPath)