+++ cmd/resim/commands/sync/experiences.go
@@ -34,7 +34,7 @@ func computeExperienceUpdates(
 		return nil, fmt.Errorf("Failed to compute system updates: %w", err)
 	}
 
-	testSuiteUpdates, err := getTestSuiteUpdates(matchedExperiencesByNewName, config.TestSuites,
+	testSuiteUpdates, err := getTestSuiteUpdates(matchedExperiencesByNewName, config.ManagedTestSuites,
 		currentState.TestSuiteIDsByName)
 	if err != nil {
 		return nil, fmt.Errorf("Failed to compute test suite updates: %w", err)
@@ -102,19 +102,19 @@ func matchExperiences(config *ExperienceSyncConfig, currentExperiencesByName map
 	matches := make(map[string]ExperienceMatch)
 
 	remainingCurrentExperiencesByID := byNameToByID(currentExperiencesByName)
-
-	for _, experience := range config.Experiences {
+	for ii := range config.Experiences {
+		experience := &config.Experiences[ii]
 		// Step 1: Attempt to match by name
 		currExp, exists := currentExperiencesByName[experience.Name]
 		if exists {
 			// If the match target has already been matched with, that's a failure
-			if _, isAvailable := remainingCurrentExperiencesByID[currExp.ExperienceID.ID]; !isAvailable {
+			if _, isAvailable := remainingCurrentExperiencesByID[*currExp.ExperienceID]; !isAvailable {
 				return nil, fmt.Errorf("Experience name collision: %s", currExp.Name)
 			}
 
 			// If it exists but its ID doesn't match a hard-coded one we provide, that's a failure
 			if currExp.ExperienceID == nil || (experience.ExperienceID != nil && *experience.ExperienceID != *currExp.ExperienceID) {
-				return nil, fmt.Errorf("Multiple experiences desire the same name: %s", experience.ExperienceID.ID)
+				return nil, fmt.Errorf("Multiple experiences desire the same name: %s", *experience.ExperienceID)
 			}
 
 			// Experience exists with the same name and should be updated
@@ -126,13 +126,13 @@ func matchExperiences(config *ExperienceSyncConfig, currentExperiencesByName map
 			if err != nil {
 				return nil, err
 			}
-			delete(remainingCurrentExperiencesByID, currExp.ExperienceID.ID)
+			delete(remainingCurrentExperiencesByID, *currExp.ExperienceID)
 			continue
 		}
 		// Step 2: Attempt to match by ID
 		if experience.ExperienceID != nil {
 			// Check if there's still an unmatched experience with this ID:
-			currExp, exists := remainingCurrentExperiencesByID[experience.ExperienceID.ID]
+			currExp, exists := remainingCurrentExperiencesByID[*experience.ExperienceID]
 			if !exists {
 				return nil, fmt.Errorf("No existing experience available with ID. This could be due to multiple configured experiences requesting the same ID: %s", *experience.ExperienceID)
 			}
@@ -144,7 +144,7 @@ func matchExperiences(config *ExperienceSyncConfig, currentExperiencesByName map
 			if err != nil {
 				return nil, err
 			}
-			delete(remainingCurrentExperiencesByID, currExp.ExperienceID.ID)
+			delete(remainingCurrentExperiencesByID, *currExp.ExperienceID)
 			continue
 		}
 
@@ -183,7 +183,7 @@ func byNameToByID(byName map[string]*Experience) map[ExperienceID]*Experience {
 	byID := make(map[ExperienceID]*Experience)
 	for _, v := range byName {
 		if v.ExperienceID != nil {
-			byID[v.ExperienceID.ID] = v
+			byID[*v.ExperienceID] = v
 		}
 	}
 	return byID